Fixed music looping
The main menu music did not loop after it ended and additionally the
check whether music was playing (after the stream ends) was not
successful.

The root cause for these problems is that the music files (.IT) were not
correctly loaded. The FModEx documentation states that music modules
will automatically be loaded as streams. However, all game sounds were
loaded using the FMOD_DEFAULT mode which specifies the flag
FMOD_LOOP_OFF. Loading the music modules by explicitly specifing the
flags FMOD_SOFTWARE and FMOD_LOOP_NORMAL seems to fix the problem.