Fixed a 64 bit arch related issue that prevented images from loading.
This is a nasty hack to load the BITMAP file positions of each animation
frame correctly.The hack comes from a stupidity I made 10 years ago to
actually serialize the pointer to the SVW_FILES structure stored inside
the SWV_HEADER structure. This was all done under 32 bit Windows where
(of course) the pointer size would be 4 bytes long. When data is loaded
back under a 64 bit machine the fread() routine will read the complete
SWV_HEADER structure in memory and map the pointer to SWV_HEADER as 64
bit -> 8 bytes. This actually "eats" some of the data further on and
breaks further fread() deserializations.

So, the solution is quite "simple". Just rewind the file cursor position
enough so we can get back on track to read the SVW_FILES array. :)