Physics fix: Added negative velocity vector on hit.
When vehicles collide the velocity of each is set to 0.0f which
effectively makes it impossible for either of them to accelerate.

We cannot add proper phsysics without rewriting most of the vehicle
class code so a small adjustment of the velocity vector (make it
negative to what it used to be) would make the vehicles repulse from
each other thus minimizing any chances to get stuck.