Added 'skip_hit_timer' that prevents collisions for a short while.
This is a debug-hack attempt to detect at which angle is the hitting
vehicle reflected. It seems normals are sometimes wrongly calculated.
This causes the vehicle to 'stick'. The sticking is actually due to
constant collision detection. Eventually velocites become 0 due to the
acceleration being decreased each frame and velocities being echanged
due to the momentum echange.