Changes in libraries linkage.
SDL shared object library is no longer copied in the game build folder
and it is no longer present in the game package. On Linux it is required
that the libsdl1.2 package is installed. On Windows the SDL.dll will be
copied in the game package by the build process.