Several fixes for Arch Linux PKGBUILD packaging
Use 'INSTALL_DATADIR' and 'INSTALL_LIBEXECDIR' in 'savagewheels.in',
so to be able to override them.

Package 'autos', 'graphics' and 'sound' dirs, only if available
in the bin/ dir. Self-extracting executables and PKGBUILD will use
that to construct the game package. Other pkg systems may just skip
it by not copying these folders in the game bin/ folder prior to
packaging.