Update handling of Linux LIBEXEC and DATADIR paths
Update the handling of the Linux LIBEXEC and DATADIR paths so that savagewheels
files are stored in ${LIBEXEC}/savagewheels and ${DATADIR}/savagewheels instead of
${LIBEXEC} and ${DATADIR}.  This change reduces the risk of filename collision with
other installed packages by placing savagewheels files in its own subdirectory.  This
change also provides a clean method of adding additional files to the LIBEXEC and
DATADIR tree if that ever becomes necessary.